+++ src/SyncClipboard.Desktop/Views/SystemSettingPage.axaml.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using SyncClipboard.Core.Clipboard;
 using SyncClipboard.Core.ViewModels;
 using System;
+using System.Diagnostics;
 using System.Threading;
 
 namespace SyncClipboard.Desktop.Views;
@@ -20,12 +21,24 @@ public partial class SystemSettingPage : UserControl
         App.Current.ExitApp();
     }
 
-    private void CopyAppDataDirPath(object? sender, Avalonia.Interactivity.RoutedEventArgs e)
+    private void CopyAppDataFolderPath(object? sender, Avalonia.Interactivity.RoutedEventArgs e)
     {
         var profile = new TextProfile(
             Core.Commons.Env.AppDataDirectory,
             App.Current.Services.GetRequiredService<IServiceProvider>()
         );
         profile.SetLocalClipboard(true, CancellationToken.None);
+    }
+
+    private void OpenDataFolderInNautilus(object? sender, Avalonia.Interactivity.RoutedEventArgs e)
+    {
+        try
+        {
+            Process.Start("nautilus", Core.Commons.Env.AppDataDirectory);
+        }
+        catch
+        {
+            App.Current.Logger.Write("Open Nautilus failed");
+        }
     }
 }